HDDS-2850. Unit test refactoring.
