+++ hadoop-ozone/recon/src/main/java/org/apache/hadoop/ozone/recon/scm/ReconIncrementalContainerReportHandler.java
@@ -98,11 +98,6 @@ public class ReconIncrementalContainerReportHandler
             replicaProto.getContainerID());
       }
     }
-
-    if (success) {
-      getContainerManager().notifyContainerReportProcessing(false, true);
-    } else {
-      getContainerManager().notifyContainerReportProcessing(false, false);
-    }
+    getContainerManager().notifyContainerReportProcessing(false, success);
   }
 }
